+++ CHANGELOG.md
@@ -30,6 +30,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#187](https://github.com/ARGOeu/argo-mon-api/pull/187) ARGO-5711 [API] Fetch monthly and daily A/R groups views for all topology levels for tenant
 - [#190](https://github.com/ARGOeu/argo-mon-api/pull/190) ARGO-5715: [API] Fetch monthly and daily A/R endpoints views for all topology levels for tenant
 - [#194](https://github.com/ARGOeu/argo-mon-api/pull/194) ARGO-5732 [API] Create Incident
+- [#200](https://github.com/ARGOeu/argo-mon-api/pull/200) ARGO-5733 [API] Change incident status / add comment
 
 ### Fix